+++ packages/node/src/plugins/LoadFileChunkLoadingRuntimeModule.ts
@@ -7,10 +7,6 @@
 
 import type {Chunk, ChunkGraph, Compiler} from 'webpack';
 import {RuntimeModule, RuntimeGlobals, Template} from 'webpack';
-
-// import RuntimeGlobals from 'webpack/lib/RuntimeGlobals';
-// import RuntimeModule from 'webpack/lib/RuntimeModule';
-// import Template from 'webpack/lib/Template';
 import {getUndoPath} from 'webpack/lib/util/identifier';
 import compileBooleanMatcher from 'webpack/lib/util/compileBooleanMatcher';
 
@@ -87,7 +83,7 @@ class ReadFileChunkLoadingRuntimeModule extends RuntimeModule {
     // remotes only use fs.readFile if we were to cache the chunks on disk after fetching - otherwise its always using http
     // so for example, if im in hostA and require(remoteb/module) --> console.log of name in runtime code will return remoteb
 
-    const {remotes, name} = this.options;
+    const {remotes = {}, name} = this.options;
     const {webpack} = this.chunkLoadingContext;
     const chunkHasJs =
       (webpack && webpack.javascript.JavascriptModulesPlugin.chunkHasJs) ||